Ass1

Documentation

data Tree a

Constructors

Leaf a

Node (Tree a) (Tree a)

unfoldr :: (s -> Maybe (a, s)) -> s -> [a]

unfoldTree :: (s -> Either a (s, s)) -> s -> Tree a

iterate :: (a -> a) -> a -> [a]

2.2.2 Iterate. [x, f x, f f x, ...] using unfoldr

map :: (a -> b) -> [a] -> [b]

2.2.2 Map. using unfoldr.

balanced :: Int -> Tree ()

2.2.2 balanced Make a balanced tree of depth n.

data SizeState

2.2.2 Record type used for clarity of code in sized

Constructors

S

Fields

depth :: Int

label :: Int

sized :: Int -> Tree Int

2.2.2 create a tree with N nodes This function will create a tree with N nodes, each of which has at least one leaf. The last one has two leaves.

class ConsumeArgs r where

2.2.5 Typeclass for consuming arguments and applying an accumulator function

Methods

consumeArgs :: (Int -> Int) -> Int -> r

count :: ConsumeArgs r => r

2.2.5 The first count starts at 0, and keeps acc constant

test :: [Int]

2.2.5 test for the first count

count' :: ConsumeArgs r => r

2.2.5 The second count starts at 0, and increments

test' :: [Int]

2.2.5 test for the second count

data PTree a

2.7.1 List type where the nodes are marked instead of the leaves, and each node has a variable amount of children.

Constructors

PRoot [PTree a]

PNode a [PTree a]

PLeaf

permutation_tree :: [a] -> PTree a

2.7.1 Generate a permutation tree for a list. List may include duplicates.

prune :: (a -> a -> Bool) -> PTree a -> PTree a

2.7.1 Prune the permutation tree, given a function.

tree2lists :: PTree a -> [[a]]

2.7.1 Turn a tree back into a list of lists

smooth_perms :: Int -> [Int] -> [[Int]]

2.7.1 Generate all smooth permutations of a list